+ optimize code by avoiding force unwrap optional.
It was crashed while using this method even though I modified this method scope to be public in my project.
It will not be crashed in the real project because this is file private but just in case, I added some optional binding for the safety reason.

Anyway, I am using this library really useful for my project.

Thanks.